Fix Log List Error: Upgrade index/count to uint16_t
- Protocol: Upgraded `total_files` and `index` in `LogListEntryMsg` from `uint8_t` to `uint16_t` (in `ecoflow_protocol.h` and `ecoflow_protocol.c` for both STM32 and ESP32) to support > 255 log files. This directly fixes the "Error loading list" issue caused by index overflow when many logs are present.
- STM32: Updated `LogManager_HandleListReq` to use `uint16_t` loop counters and optimized throttling (delay every 10 packets instead of every packet) to improve listing speed while maintaining UI responsiveness.
- ESP32: Updated `Stm32Serial` to handle `uint16_t` log list indices and increased the `getLogList` timeout to 20 seconds to accommodate larger directories.
- Verified compilation on both platforms.

+++ EcoflowESP32/lib/EcoFlowComm/ecoflow_protocol.h
@@ -331,8 +331,8 @@ typedef struct {
 } LogListRespEntry;
 
 typedef struct {
-    uint8_t total_files;
-    uint8_t index;
+    uint16_t total_files;
+    uint16_t index;
     uint32_t size;
     char name[32];
 } LogListEntryMsg;
@@ -416,8 +416,8 @@ int pack_ota_apply_message(uint8_t *buffer);
 
 // Log API
 int pack_log_list_req_message(uint8_t *buffer);
-int pack_log_list_resp_message(uint8_t *buffer, uint8_t total, uint8_t index, uint32_t size, const char* name);
-int unpack_log_list_resp_message(const uint8_t *buffer, uint8_t *total, uint8_t *index, uint32_t *size, char* name);
+int pack_log_list_resp_message(uint8_t *buffer, uint16_t total, uint16_t index, uint32_t size, const char* name);
+int unpack_log_list_resp_message(const uint8_t *buffer, uint16_t *total, uint16_t *index, uint32_t *size, char* name);
 
 int pack_log_download_req_message(uint8_t *buffer, const char* name);
 int unpack_log_download_req_message(const uint8_t *buffer, char* name);
